BizApp/analytics/requirements.txt
pandas
numpy
scikit-learn
xgboost
shap
pyodbc
sqlalchemy
python-dotenv
kaggle
fastapi
uvicorn[standard]
pydantic
joblib